requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.8
pillow==12.1.1
PyJWT==2.12.1
sqlparse==0.5.5
whitenoise==6.12.0
